Split process into header and source, removed some redundant comments and moved as many includes as possible to the source
