+++ applications/GeoMechanicsApplication/custom_processes/apply_scalar_constraint_table_process.cpp
@@ -11,7 +11,7 @@
 //                   Marjan Fathian
 //
 #include "apply_scalar_constraint_table_process.h"
-#include "apply_component_table_process.hpp"
+#include "apply_component_table_process.h"
 #include "apply_constant_interpolate_line_pressure_process.hpp"
 #include "apply_constant_phreatic_line_pressure_process.hpp"
 #include "apply_constant_phreatic_surface_pressure_process.hpp"
