+++ applications/GeoMechanicsApplication/custom_python/add_custom_processes_to_python.cpp
@@ -23,7 +23,7 @@
 #include "custom_processes/apply_boundary_phreatic_line_pressure_table_process.hpp"
 #include "custom_processes/apply_boundary_phreatic_surface_pressure_table_process.hpp"
 #include "custom_processes/apply_c_phi_reduction_process.h"
-#include "custom_processes/apply_component_table_process.hpp"
+#include "custom_processes/apply_component_table_process.h"
 #include "custom_processes/apply_constant_boundary_hydrostatic_pressure_process.hpp"
 #include "custom_processes/apply_constant_boundary_phreatic_line_pressure_process.hpp"
 #include "custom_processes/apply_constant_boundary_phreatic_surface_pressure_process.hpp"
