+++ applications/GeoMechanicsApplication/custom_workflows/dgeoflow.h
@@ -33,7 +33,7 @@
 #include "solving_strategies/builder_and_solvers/residualbased_block_builder_and_solver.h"
 
 // The strategies to test
-#include <custom_processes/apply_component_table_process.hpp>
+#include <custom_processes/apply_component_table_process.h>
 #include <custom_processes/apply_constant_hydrostatic_pressure_process.hpp>
 #include <linear_solvers/skyline_lu_factorization_solver.h>
